Fix English FAQ rich result schema
