+++ src/app/[locale]/academic/english/layout.tsx
@@ -1,11 +1,5 @@
 import { Metadata } from 'next'
-import BreadcrumbSchema from '@/components/schema/BreadcrumbSchema'
-import FAQSchema from '@/components/schema/FAQSchema'
-import { ENGLISH_COURSE_MERGED_FAQ_JSONLD } from '@/lib/schema/course-hub-jsonld-faqs'
 import { generateMetadataFromPath } from '@/lib/seo/metadata'
-import { generateCourseSchema } from '@/lib/seo/structuredData'
-import { absoluteSiteUrl } from '@/lib/publicPath'
-import { getCanonicalSiteUrl } from '@/lib/seo/siteUrl'
 
 export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }): Promise<Metadata> {
   const { locale } = await params
@@ -15,63 +9,8 @@ export async function generateMetadata({ params }: { params: Promise<{ locale: s
 
 export default async function EnglishCoursesLayout({
   children,
-  params,
 }: {
   children: React.ReactNode
-  params: Promise<{ locale: string }>
 }) {
-  const { locale } = await params
-  const baseUrl = getCanonicalSiteUrl()
-  
-  const courseSchema = generateCourseSchema({
-    name: "English Language Arts Courses - Reading, Writing & Grammar | English Tutoring Dublin CA",
-    description:
-      "English and ELA tutoring for grades 1–12 in Dublin, CA. Reading, writing, and grammar in small groups. Book a free assessment.",
-    provider: "GrowWise",
-    courseCode: "ELA-K12",
-    educationalLevel: "Grades 1-12",
-    teaches: [
-      "Reading Comprehension",
-      "Vocabulary Development",
-      "Grammar & Mechanics",
-      "Essay Writing",
-      "Creative Writing",
-      "English Language Arts",
-      "Literary Analysis",
-      "Writing Skills"
-    ],
-    coursePrerequisites: "Placement assessment recommended to determine appropriate level",
-    url: absoluteSiteUrl('/academic/english', locale, baseUrl),
-    image: `${baseUrl}/assets/growwise-logo.png`,
-    offers: {
-      price: "35",
-      priceCurrency: "USD",
-      availability: "https://schema.org/InStock",
-      url: absoluteSiteUrl('/enroll', locale, baseUrl),
-    }
-  })
-
-  return (
-    <>
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: JSON.stringify(courseSchema) }}
-      />
-      <BreadcrumbSchema
-        items={[
-          { name: 'Home', url: absoluteSiteUrl('/', locale, baseUrl) },
-          {
-            name: 'Academic Programs',
-            url: absoluteSiteUrl('/academic', locale, baseUrl),
-          },
-          {
-            name: 'English Courses',
-            url: absoluteSiteUrl('/academic/english', locale, baseUrl),
-          },
-        ]}
-      />
-      <FAQSchema faqs={ENGLISH_COURSE_MERGED_FAQ_JSONLD} />
-      {children}
-    </>
-  )
+  return <>{children}</>
 }
