+++ src/lib/__tests__/resources-internal-links.test.ts
@@ -73,7 +73,7 @@ describe('resources internal links', () => {
   });
 
   it('links courses/english to reading-fluency-vs-comprehension', () => {
-    const source = readComponent('app/[locale]/academic/english/page.tsx');
+    const source = readComponent('components/EnglishCoursesPage.tsx');
     expect(source).toContain('/resources/reading-fluency-vs-comprehension');
   });
