+++ src/lib/seo/__tests__/seo-jsonld-route-audit.test.ts
@@ -69,7 +69,6 @@ const ROUTE_JSON_LD_AUDIT: Array<{ route: string; schemas: unknown[] }> = [
     schemas: [
       { '@type': 'Course' },
       { '@type': 'BreadcrumbList' },
-      generateFAQPageSchema([]),
     ],
   },
   {
@@ -130,7 +129,6 @@ describe('SEO JSON-LD route audit', () => {
         '/steam',
         '/programs',
         '/academic/math',
-        '/academic/english',
         '/dublin-ca',
         '/resources/tutoring-dublin-ca',
         '/resources/careless-math-mistakes',
